Refactor pkg/controller and fix typo on CRD name
