+++ examples/controller/variant.crds.yaml
@@ -17,7 +17,7 @@ spec:
 apiVersion: apiextensions.k8s.io/v1beta1
 kind: CustomResourceDefinition
 metadata:
-  name: reconcilations.core.variant.run
+  name: reconciliations.core.variant.run
 spec:
   group: core.variant.run
   versions:
@@ -25,7 +25,7 @@ spec:
     served: true
     storage: true
   names:
-    kind: Reconcilation
-    plural: reconcilations
-    singular: reconcilation
+    kind: Reconciliation
+    plural: reconciliations
+    singular: reconciliation
   scope: Namespaced
